requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
python-dotenv
psycopg2